Removed show() from unit tests, moved plot tests to test_plots.py
